Rename server source folders from 'java' to 'kotlin', remove unnecessary project files, refactor code
