--- goty-server/lombok.config
@@ -1,3 +1,0 @@
-# This file is generated by the 'io.freefair.lombok' Gradle plugin
-config.stopBubbling = true
-lombok.anyConstructor.addConstructorProperties = true
