+++ goty-server/src/main/kotlin/com/aleinin/goty/GameOfTheYearApplication.kt
