+++ goty-server/src/main/kotlin/com/aleinin/goty/elasticsearch/ElasticsearchProperties.kt
