+++ goty-server/src/main/kotlin/com/aleinin/goty/game/GameSearchRequest.kt
