+++ goty-server/src/main/kotlin/com/aleinin/goty/game/GameSearchResponse.kt
