+++ goty-server/src/main/kotlin/com/aleinin/goty/objectmapper/ObjectMapperConfiguration.kt
