+++ goty-server/src/main/kotlin/com/aleinin/goty/submit/SubmissionProperties.kt
