+++ goty-server/src/test/kotlin/com/aleinin/goty/GameOfTheYearApplicationTests.kt
